+++ aiter/jit/core.py
@@ -27,6 +27,10 @@ from torch_guard import torch_compile_guard  # noqa: E402
 
 AITER_REBUILD = int(os.environ.get("AITER_REBUILD", "0"))
 ENABLE_CK = int(os.environ.get("ENABLE_CK", "1")) != 0
+# Runtime switch for the optional FlyDSL backend, mirroring ENABLE_CK. Set
+# ENABLE_FLYDSL=0 to fully disable FlyDSL kernel imports/usage (e.g. when the
+# installed flydsl version is incompatible with the current source).
+ENABLE_FLYDSL = int(os.environ.get("ENABLE_FLYDSL", "1")) != 0
 
 
 def is_experimental_enabled() -> bool:
